+++ android/anc/src/main/java/org/smartregister/fhircore/anc/data/AncPaginatedDataSource.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.smartregister.fhircore.anc.ui.anccare
+package org.smartregister.fhircore.anc.data
 
 import com.google.android.fhir.FhirEngine
 import com.google.android.fhir.search.Search
@@ -22,7 +22,7 @@ import com.google.android.fhir.search.StringFilterModifier
 import org.hl7.fhir.r4.model.Patient
 import org.smartregister.fhircore.engine.data.domain.util.DomainMapper
 import org.smartregister.fhircore.engine.data.domain.util.PaginatedDataSource
-import org.smartregister.fhircore.engine.data.local.repository.patient.model.AncItem
+import org.smartregister.fhircore.anc.data.model.AncItem
 
 class AncPaginatedDataSource(
   val fhirEngine: FhirEngine,
