+++ android/anc/src/main/java/org/smartregister/fhircore/anc/data/AncRepository.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.smartregister.fhircore.anc.ui.anccare
+package org.smartregister.fhircore.anc.data
 
 import com.google.android.fhir.FhirEngine
 import com.google.android.fhir.search.Order
@@ -24,14 +24,14 @@ import kotlinx.coroutines.withContext
 import org.hl7.fhir.r4.model.Patient
 import org.smartregister.fhircore.engine.data.domain.util.DomainMapper
 import org.smartregister.fhircore.engine.data.domain.util.RegisterRepository
-import org.smartregister.fhircore.engine.data.local.repository.patient.model.AncItem
+import org.smartregister.fhircore.anc.data.model.AncItem
 import org.smartregister.fhircore.engine.util.DefaultDispatcherProvider
 import org.smartregister.fhircore.engine.util.DispatcherProvider
 
 class AncRepository(
-  override val fhirEngine: FhirEngine,
-  override val domainMapper: DomainMapper<Patient, AncItem>,
-  private val dispatcherProvider: DispatcherProvider = DefaultDispatcherProvider
+    override val fhirEngine: FhirEngine,
+    override val domainMapper: DomainMapper<Patient, AncItem>,
+    private val dispatcherProvider: DispatcherProvider = DefaultDispatcherProvider
 ) : RegisterRepository<Patient, AncItem> {
 
   override val defaultPageSize: Int
