+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/login/LoginActivity.kt
@@ -20,7 +20,7 @@ import android.content.Intent
 import android.os.Bundle
 import org.smartregister.fhircore.anc.BuildConfig
 import org.smartregister.fhircore.anc.R
-import org.smartregister.fhircore.anc.ui.anccare.AncRegisterActivity
+import org.smartregister.fhircore.anc.ui.anccare.register.AncRegisterActivity
 import org.smartregister.fhircore.engine.configuration.view.loginViewConfigurationOf
 import org.smartregister.fhircore.engine.ui.login.BaseLoginActivity
